+++ tests/faces/test_bsplineface3d.py
@@ -2,7 +2,7 @@ import unittest
 import os
 from time import perf_counter
 import volmdlr
-from volmdlr import edges, surfaces, wires, faces
+from volmdlr import edges, surfaces, wires, faces, core
 from volmdlr.models.bspline_surfaces import bspline_surface_1
 
 
@@ -51,6 +51,14 @@ class TestBSplineFace3D(unittest.TestCase):
         face = faces.BSplineFace3D.from_contours3d(surface, [contour3d])
         self.assertTrue(face.surface2d.outer_contour.is_ordered(1e-4))
         self.assertAlmostEqual(face.surface2d.area(), 0.9962228649263708, 2)
+
+        surface = surfaces.BSplineSurface3D.load_from_file(
+            os.path.join(folder, "bsplinesurface_bsplineface_with_openned_contour.json"))
+        contours3d = core.VolumeModel.load_from_file(
+            os.path.join(folder, "bsplinesurface_bsplineface_with_openned_contour_contours.json")).primitives
+        face = faces.BSplineFace3D.from_contours3d(surface, contours3d)
+        self.assertAlmostEqual(face.surface2d.area(),0.4261703133157918, 2)
+
 
     def test_neutral_fiber(self):
         face = faces.BSplineFace3D.load_from_file(os.path.join(folder, "test_neutral_fiber.json"))